Move CudaKernel definition into a separate header file.
